Makes the car go and stop when sensor triggers
The code necessary for the Basic Obstacle Avoidance assignment. We have written more advanced obstacle avoidance already, which is on the Advanced Obstacle Avoidance branch. Not sure if this needs to be merged into master for the completion of the assignment, but I need to provide a pull request to turn in.